+++ lua/obsidian/init.lua
@@ -115,6 +115,18 @@ obsidian.setup = function(opts)
 
   local group = vim.api.nvim_create_augroup("obsidian_setup", { clear = true })
 
+  -- wrapper for creating autocmd events
+  ---@param pattern string
+  ---@param buf integer
+  local function exec_autocmds(pattern, buf)
+    vim.api.nvim_exec_autocmds("User", {
+      pattern = pattern,
+      data = {
+        note = require("obsidian.note").from_buffer(buf),
+      },
+    })
+  end
+
   -- Complete setup and update workspace (if needed) when entering a markdown buffer.
   vim.api.nvim_create_autocmd({ "BufEnter" }, {
     group = group,
@@ -153,8 +165,10 @@ obsidian.setup = function(opts)
 
       -- Run enter-note callback.
       client.callback_manager:enter_note(function()
-        return obsidian.Note.from_buffer(ev.bufnr)
+        return obsidian.Note.from_buffer(ev.buf)
       end)
+
+      exec_autocmds("ObsidianNoteEnter", ev.buf)
     end,
   })
 
@@ -175,8 +189,10 @@ obsidian.setup = function(opts)
 
       -- Run leave-note callback.
       client.callback_manager:leave_note(function()
-        return obsidian.Note.from_buffer(ev.bufnr)
+        return obsidian.Note.from_buffer(ev.buf)
       end)
+
+      exec_autocmds("ObsidianNoteLeave", ev.buf)
     end,
   })
 
@@ -205,10 +221,33 @@ obsidian.setup = function(opts)
       -- Run pre-write-note callback.
       client.callback_manager:pre_write_note(note)
 
+      exec_autocmds("ObsidianNoteWritePre", ev.buf)
+
       -- Update buffer with new frontmatter.
       if client:update_frontmatter(note, bufnr) then
         log.info "Updated frontmatter"
       end
+    end,
+  })
+
+  vim.api.nvim_create_autocmd({ "BufWritePost" }, {
+    group = group,
+    pattern = "*.md",
+    callback = function(ev)
+      local buf_dir = vim.fs.dirname(ev.match)
+
+      -- Check if we're in a workspace.
+      local workspace = obsidian.Workspace.get_workspace_for_dir(buf_dir, client.opts.workspaces)
+      if not workspace then
+        return
+      end
+
+      -- Check if current buffer is actually a note within the workspace.
+      if not client:path_is_note(ev.match, workspace) then
+        return
+      end
+
+      exec_autocmds("ObsidianNoteWritePost", ev.buf)
     end,
   })